// src/web/src/api/models/childGuardianVM.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * KDVManager CRM API
 * OpenAPI spec version: v1
 */
import type { GuardianRelationshipType } from "./guardianRelationshipType";

export type ChildGuardianVM = {
  guardianId?: string;
  fullName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  relationshipType?: GuardianRelationshipType;
  isPrimaryContact?: boolean;
  isEmergencyContact?: boolean;
};